refactor: rename cicd_pipeline_commits.repo to repo_url
